Reorganise modules: list_parser -> unpack_list, make_meta
